Retrieves variant-surrounding genes with calculated statistics for causal gene prioritization.

This function retrieves all calculated prioritizing scores for surrounding genes of a specific variant based on the Open Target Genetics locus-to-gene (L2G) ML scoring pipeline. It provides detailed insights into the relationship between genetic variants and genes, allowing users to explore the impact of variants on gene expression, colocalization scores, chromatin interactions, and predicted functional effects. The function returns the information in a structured format, making it easier to analyze and interpret the results.

Usage

genesForVariant(variant_id)

Arguments

variant_id

A character string specifying the ID of the variant for which to fetch gene information.

Value

A list with the following components:

v2g: A data frame with all variant-to-gene information with the following data structure: - gene.symbol: character - variant: character - overallScore: numeric - gene.id: character

tssd: A data frame with all details on colocalization scores effect of variant in expression of the genes across tissues with the following data structure: - gene.symbol: character - variant: character - typeId: character - sourceId: character - aggregatedScore: numeric - tissues_distance: integer - tissues_score: numeric - tissues_quantile: numeric - tissues_id: character - tissues_name: character

qtls: List of QTL associations between genes and variants across analyzed tissues with the following data structure: - gene.symbol: character - variant: character - typeId: character - aggregatedScore: numeric - tissues_quantile: numeric - tissues_beta: numeric - tissues_pval: numeric - tissues_id: character - tissues_name: character

chromatin: A data frame including all information on chromatin interactions effect involving genes and variants with the following data structure: - gene.symbol: character - variant: character - typeId: character - sourceId: character - aggregatedScore: numeric - tissues_quantile: numeric - tissues_score: numeric - tissues_id: character - tissues_name: character

functionalpred: A data frame including predicted functional effects of variants on genes across tissues with the following data structure: - gene.symbol: character - variant: character - typeId: character - sourceId: character - aggregatedScore: numeric - tissues_maxEffectLabel: character - tissues_maxEffectScore: numeric - tissues_id: character - tissues_name: character

Examples

if (FALSE) { # \dontrun{
  result <- genesForVariant(variant_id = "1_154453788_C_T")
  result <- genesForVariant(variant_id = "rs4129267")
  print(result)
} # }