Retrieve calculated GWAS colocalisation data

This function retrieves colocalisation data for a specific variant from a study with other GWAS studies. It returns a data frame of the studies that colocalise with the input variant and study, including details on the study and reported trait, index variant, and calculated coloc method (see Ref. below) outputs.

Usage

gwasColocalisation(study_id, variant_id)

Arguments

study_id

Character: Open Target Genetics generated ID for the GWAS study.

variant_id

Character: Open Target Genetics generated ID for the variant (CHRPOSITION_REFALLELE_ALTALLELE or rsID).

Value

Returns a data frame of the studies that colocalise with the input variant and study. The table consists of the following data structure:

study.studyId: Character vector. Study identifier.

study.traitReported: Character vector. Reported trait associated with the colocalisation.

study.traitCategory: Character vector. Trait category.

indexVariant.id: Character vector. Index variant identifier.

indexVariant.position: Integer vector. Index variant position.

indexVariant.chromosome: Character vector. Index variant chromosome.

indexVariant.rsId: Character vector. Index variant rsID.

beta: Numeric vector. Beta value associated with the colocalisation.

h3: Numeric vector. H3 value associated with the colocalisation.

h4: Numeric vector. H4 value associated with the colocalisation.

log2h4h3: Numeric vector. Log2 ratio of H4 to H3 values.

References

Giambartolomei, Claudia et al. “Bayesian test for colocalisation between pairs of genetic association studies using summary statistics.” PLoS genetics vol. 10,5 e1004383. 15 May. 2014, doi:10.1371/journal.pgen.1004383

Examples

if (FALSE) { # \dontrun{
colocalisation_data <- gwasColocalisation(study_id = "GCST90002357", variant_id = "1_154119580_C_A")
colocalisation_data <- gwasColocalisation(study_id = "GCST90002357", variant_id = "rs2494663")
} # }