Retrieve GWAS summary statistics to create a Manhattan plot

This function retrieves summary statistics for a given GWAS study ID, which are used to generate a Manhattan plot. The Manhattan plot is a graphical representation of genetic association studies, particularly in genome-wide association studies (GWAS). It displays the results of statistical associations between genetic variants and a trait or disease of interest across the genome. This function returns a data frame of the underlying data, which can be used to recreate the Manhattan plot using the plot_manhattan() function, or for custom plots and downstream analysis.

Usage

manhattan(study_id, pageindex = 0, pagesize = 100)

Arguments

study_id

Character: Open Targets Genetics generated ID for the GWAS study.

pageindex

Int: Index of the current page (pagination index >= 0).

pagesize

Int: Number of records in a page (pagination size > 0).

Value

Returns a data frame containing the Manhattan associations for the input study ID. The table consists of the following columns:

pval_mantissa: Numeric vector. Mantissa of the p-value.

pval_exponent: Integer vector. Exponent of the p-value.

credible_set_size: Integer vector. Size of the credible set.

ld_set_size: Integer vector. Size of the LD set.

total_set_size: Integer vector. Total size of the set.

pval: Numeric vector. P-value.

odds_ratio: Logical vector. Odds ratio.

odds_ratio_ci_lower: Logical vector. Lower confidence interval of the odds ratio.

odds_ratio_ci_upper: Logical vector. Upper confidence interval of the odds ratio.

beta: Numeric vector. Beta value.

beta_ci_lower: Numeric vector. Lower confidence interval of the beta value.

beta_ci_upper: Numeric vector. Upper confidence interval of the beta value.

direction: Character vector. Direction of the effect.

best_genes_score: Numeric vector. Score of the best genes.

best_genes_gene_id: Character vector. Gene ID of the best genes.

best_genes_gene_symbol: Character vector. Gene symbol of the best genes.

best_coloc_genes_score: Numeric vector. Score of the best colocated genes.

best_coloc_genes_gene_id: Character vector. Gene ID of the best colocated genes.

best_coloc_genes_gene_symbol: Character vector. Gene symbol of the best colocated genes.

best_locus2genes_score: Numeric vector. Score of the best locus-to-genes.

best_locus2genes_gene_id: Character vector. Gene ID of the best locus-to-genes.

best_locus2genes_gene_symbol: Character vector. Gene symbol of the best locus-to-genes.

variant_id: Character vector. Variant ID.

variant_position: Integer vector. Variant position.

variant_chromosome: Character vector. Variant chromosome.

variant_rs_id: Character vector. Variant rsID.

Examples

if (FALSE) { # \dontrun{
result <- manhattan(study_id = "GCST90002357")
result <- manhattan(study_id = "GCST90002357", pageindex = 2, pagesize = 50)
} # }