Retrieve QTL colocalisation results of a variant

The colocalisation analysis in Open Target Genetics is performed using the coloc method (Giambartolomei et al., 2014). Coloc is a Bayesian method which, for two traits, integrates evidence over all variants at a locus to evaluate the following hypotheses: - H0: No association with either trait - H1: Association with trait 1, not with trait 2 - H2: Association with trait 2, not with trait 1 - H3: Association with trait 1 and trait 2, two independent SNPs - H4: Association with trait 1 and trait 2, one shared SNP This analysis tests whether two independent associations at the same locus are consistent with having a shared causal variant. Colocalisation of two independent associations from two GWAS studies may suggest a shared causal mechanism.

Usage

qtlColocalisationVariantQuery(study_id, variant_id)

Arguments

study_id

Character: Study ID(s) generated by Open Targets Genetics (e.g GCST90002357).

variant_id

Character: generated ID for variants by Open Targets Genetics (e.g. 1_154119580_C_A) or rsId (rs2494663).

Value

Returns a data frame of the colocalisation information for a lead variant in a specific study. The output is a tidy data frame with the following data structure:

qtlStudyName: Character vector. QTL study name.

phenotypeId: Character vector. Phenotype ID.

gene.id: Character vector. Gene ID.

gene.symbol: Character vector. Gene symbol.

name: Character vector. Tissue name.

indexVariant.id: Character vector. Index variant ID.

indexVariant.rsId: Character vector. Index variant rsID.

beta: Numeric. Beta value.

h4: Numeric. h4 value.

h3: Numeric. h3 value.

log2h4h3: Numeric. Log2(h4/h3) value.

Examples

if (FALSE) {
result <- qtlColocalisationVariantQuery(study_id = "GCST90002357", variant_id = "1_154119580_C_A")
result <- qtlColocalisationVariantQuery(study_id = "GCST90002357", variant_id = "rs2494663")
}