Retrieve study summary information.

For a given study id, this function returns a data frame of relevant information about the GWAS study, such as PubMed ID, studied trait EFO ID, case/control size, etc.

Usage

studyInfo(study_id)

Arguments

study_id

Character: Study ID(s) generated by Open Targets Genetics (e.g GCST90002357).

Value

Returns a data frame (in tibble format) containing the summary iformation about a GWAS study. The data frame has the following data structure:

studyId: Character. Study ID.

traitReported: Character. Reported trait.

source: Character. Source.

traitEfos: Character. Trait EFO ID.

pmid: Character. PubMed ID.

pubDate: Character. Publication date.

pubJournal: Character. Publication journal.

pubTitle: Character. Publication title.

pubAuthor: Character. Publication author.

hasSumstats: Character. Indicates if the study has summary statistics.

ancestryInitial: Character. Initial ancestry.

nInitial: Character. Initial sample size.

nReplication: Character. Replication sample size.

traitCategory: Character. Trait category.

numAssocLoci: Character. Number of associated loci.

nTotal: Character. Total sample size.

Examples

if (FALSE) { # \dontrun{
result <- studyInfo(study_id = "GCST90002357")
} # }