Retrieve the locus-to-gene (L2G) data table for loci genes.

This function fetches the locus-to-gene (L2G) pipeline summary data table for the neighboring genes of a variant in a GWAS study.

Usage

studyLocus2GeneTable(study_id, variant_id)

Arguments

study_id

Character: Study ID(s) generated by Open Targets Genetics (e.g GCST90002357).

variant_id

Character: generated ID for variants by Open Targets Genetics (e.g. 1_154119580_C_A) or rsId (rs2494663).

Value

Returns a data frame with the summary statistics of the study and a data table containing various calculated scores and features for any lead variant. The output table has the following data structure:

studyId: Character. Study ID.

variant.id: Character. Variant ID.

variant.rsId: Character. Variant rsID.

yProbaDistance: Numeric. Distance score.

yProbaModel: Numeric. Model score.

yProbaMolecularQTL: Numeric. Molecular QTL score.

yProbaPathogenicity: Numeric. Pathogenicity score.

yProbaInteraction: Numeric. Interaction score.

hasColoc: Logical. Indicates if colocalization data is available.

distanceToLocus: Numeric. Distance to the locus.

gene.id: Character. Gene ID.

gene.symbol: Character. Gene symbol.

Examples

if (FALSE) { # \dontrun{
result <- studyLocus2GeneTable(study_id = "GCST90002357", variant_id = "1_154119580_C_A")
result <- studyLocus2GeneTable(study_id = "GCST90002357", variant_id = "rs2494663")
} # }