Retrieves all variants for a study.

For an input study ID, this function returns information of all variants across associated loci. The output also includes information about the associated genes within the each loci.

Usage

studyVariants(study_id)

Arguments

study_id

Character: Study ID(s) generated by Open Targets Genetics (e.g GCST90002357).

Value

Returns a list of two data frames.

the first data frame (tibble format) includes the loci data frame with following data structure:

variant.id: Character. Variant ID.

pval: Numeric. P-value.

variant.nearestCodingGene.symbol: Character. Symbol of the nearest coding gene to the variant.

variant.rsId: Character. Variant rsID.

variant.chromosome: Character. Chromosome of the variant.

variant.position: Integer. Position of the variant.

variant.nearestCodingGeneDistance: Integer. Distance to the nearest coding gene.

credibleSetSize: Integer. Size of the credible set.

ldSetSize: Integer. Size of the LD set.

oddsRatio: Numeric. Odds ratio.

beta: Numeric. Beta value.

The second data frame includes gene information with following data structure:

score: Numeric. Gene score.

gene.id: Character. Gene ID.

gene.symbol: Character. Gene symbol.

Examples

if (FALSE) {
result <- studyVariants(study_id = "GCST003155")
}