Retrieve GWAS colocalisation data for a genomic region .

By providing a genomic region (chromosome name with start and end position), this function returns information about colocalisation between GWAS studies and associated loci within a specified genomic region. It provides details on the studies that have at least one overlapping associated locus within the region, allowing for the assessment of potential shared causal variants. The query output includes data such as the study identifiers, traits, loci information, and other relevant attributes.

Usage

gwasColocalisationForRegion(chromosome, start, end)

Arguments

chromosome

String: Chromosome number as a string.

start

Long: Start position of the specified chromosome.

end

Long: End position of the specified chromosome.

Value

Returns a data frame with the following columns:

leftVariant.id: Character vector. ID of the left variant.

leftVariant.position: Integer vector. Position of the left variant.

leftVariant.chromosome: Character vector. Chromosome of the left variant.

leftVariant.rsId: Character vector. rsID of the left variant.

leftStudy.studyId: Character vector. Study identifier for the left study.

leftStudy.traitReported: Character vector. Reported trait associated with the colocalisation in the left study.

leftStudy.traitCategory: Character vector. Category of the reported trait in the left study.

rightVariant.id: Character vector. ID of the right variant.

rightVariant.position: Integer vector. Position of the right variant.

rightVariant.chromosome: Character vector. Chromosome of the right variant.

rightVariant.rsId: Character vector. rsID of the right variant.

rightStudy.studyId: Character vector. Study identifier for the right study.

rightStudy.traitReported: Character vector. Reported trait associated with the colocalisation in the right study.

rightStudy.traitCategory: Character vector. Category of the reported trait in the right study.

h3: Numeric vector. H3 value associated with the colocalisation.

h4: Numeric vector. H4 value associated with the colocalisation.

log2h4h3: Numeric vector. Log2 ratio of H4 to H3 values associated with the colocalisation.

References

- Giambartolomei, Claudia et al. “Bayesian test for colocalisation between pairs of genetic association studies using summary statistics.” PLoS genetics vol. 10,5 e1004383. 15 May. 2014, doi:10.1371/journal.pgen.1004383

Examples

if (FALSE) {
result <- gwasColocalisationForRegion(chromosome = "1", start = 153992685, end = 154155116)
}