Retrieves overlap info for a study and a list of studies

For an input study ID and a list of other study IDs, this function returns two elements. One contains the overlap information in a table format, and the other element is the variant intersection set, representing an overlap between two variants of the two given studies.

Usage

overlapInfoForStudy(study_id, study_ids = list())

Arguments

study_id

Character: Study ID(s) generated by Open Targets (e.g GCST90002357).

study_ids

Character: generated ID for variants by Open Targets (e.g. 1_154119580_C_A) or rsId (rs2494663).

Value

A list containing a data frame of overlap information and the variant intersection set. The overlap information table (overlap_info) consists of the following columns:

studyId: Character vector. Study ID.

traitReported: Character vector. Reported trait.

traitCategory: Character vector. Trait category.

variantIdA: Character vector. Variant ID from study A.

variantIdB: Character vector. Variant ID from study B.

overlapAB: Integer vector. Number of overlaps between variants A and B.

distinctA: Integer vector. Number of distinct variants in study A.

distinctB: Integer vector. Number of distinct variants in study B.

study.studyId: Character vector. Study ID from study list.

study.traitReported: Character vector. Reported trait from study list.

study.traitCategory: Character vector. Trait category from study list.

The variant intersection set (variant_intersection_set) is a character vector representing the intersection of variants.

Examples

if (FALSE) { # \dontrun{
result <- overlapInfoForStudy(study_id = "GCST90002357",
 study_ids = list("GCST90025975", "GCST90025962"))
} # }