Scatter plot of colocalisations for gene and reported traits.

Generates a scatter plot using the results from colocalisationsForGene() function as an input. The reported trait in each study are shown on the x-axis and plotted against their corresponding -log2(H4/H3) values on the y-axis, indicating the evidence of colocalisation between the molecular QTLs reported in each study and the explored gene. The molecular QTLs are mapped to the colors of the points. If the results of colocalisationsForGene() includes the data for multiple genes, they will be plotted in separate panels.

Usage

plot_coloc(data, biobank = FALSE)

Arguments

data

Data Frame: result of colocalisationsForGene function in data frame format, contacting the phewas information for a variant id

biobank

Logical: TRUE and FALSE variable, with the default value of FALSE which will keep the data that are from UKbioBank beside the published GWAS data. In case, this parameter is set to TRUE, only UKbioBank data will be kept which has not been published.

Value

A horizontal bar plot for colocalisation of information.

Examples

if (FALSE) {
plot_out <- colocalisationsForGene(genes = "ENSG00000169174") %>%
          plot_coloc(biobank = FALSE)
plot_out <- colocalisationsForGene(genes = "PCSK9") %>%
          plot_coloc(biobank = TRUE)
}