Retrieve calculated QTL summary statistics for credible variant set.

In Open Targets Genetics, the lead variants are expanded into a more comprehensive set of candidate causal variants referred to as the tag variants. This function retrieves calculated summary statistics for tag variants included in a lead variant colocalization analysis for a given study (which links a top loci with a trait). The user can filter the results by desired biofeature (e.g tissue, cell type,...) the function obtains tag variant information.

Usage

qtlCredibleSet(study_id, variant_id, gene, biofeature)

Arguments

study_id

Character: Study ID(s) generated by Open Targets Genetics (e.g GCST90002357).

variant_id

Character: generated ID for variants by Open Targets Genetics (e.g. 1_154119580_C_A) or rsId (rs2494663).

gene

Character: Gene ENSEMBL ID (e.g. ENSG00000169174) or gene symbol (e.g. PCSK9).

biofeature

Character: Represents either a tissue, cell type, aggregation type, protein type, etc.

Value

Returns a data frame of results from the QTL credible set of variants consisting of the following columns:

tagVariant.id: Character vector. Tag variant ID.

tagVariant.rsId: Character vector. Tag variant rsID.

pval: Numeric. P-value.

se: Numeric. Standard error.

beta: Numeric. Beta value.

postProb: Numeric. Posterior probability.

MultisignalMethod: Character vector. Multisignal method.

logABF: Numeric. Logarithm of approximate Bayes factor.

is95: Logical. Indicates if the variant has a 95

is99: Logical. Indicates if the variant has a 99

Examples

if (FALSE) {
result <- qtlCredibleSet(study_id = "Braineac2", variant_id = "1_55053079_C_T",
    gene = "ENSG00000169174", biofeature = "SUBSTANTIA_NIGRA")
result <- qtlCredibleSet(study_id = "Braineac2", variant_id = "rs7552841",
    gene = "PCSK9", biofeature = "SUBSTANTIA_NIGRA")
}