Retrieves information about a variant.

For a given variant ID, this function retrieves information about the variant, including its chromosome, position, reference allele, alternative allele, rsID, nearest gene, most severe consequence, and allele frequencies in different populations from gnomAD databse. The Genome Aggregation Database (gnomAD) is a resource developed by an international coalition of investigators, with the goal of aggregating and harmonizing both exome and genome sequencing data from a wide variety of large-scale sequencing projects, and making summary data available for the wider scientific community (see the reference).

Usage

variantInfo(variant_id)

Arguments

variant_id

Character: generated ID for variants by Open Targets Genetics (e.g. 1_154119580_C_A) or rsId (rs2494663).

Value

Returns a data frame (in tibble format) containing information about the variant. The data frame has the following structure:

chromosome: Character. Chromosome of the variant.

position: Integer. Position of the variant.

refAllele: Character. Reference allele.

altAllele: Character. Alternative allele.

rsId: Character. Variant rsID.

chromosomeB37: Character. Chromosome of the variant in build 37 coordinates.

positionB37: Integer. Position of the variant in build 37 coordinates.

id: Character. Variant ID.

nearestGene.id: Character. ID of the nearest gene to the variant.

nearestGene.symbol: Character. Symbol of the nearest gene to the variant.

nearestGeneDistance: Integer. Distance to the nearest gene.

nearestCodingGene.id: Character. ID of the nearest coding gene to the variant.

nearestCodingGene.symbol: Character. Symbol of the nearest coding gene to the variant.

nearestCodingGeneDistance: Integer. Distance to the nearest coding gene.

mostSevereConsequence: Character. Most severe consequence of the variant.

caddRaw: Numeric. CADD raw score.

caddPhred: Numeric. CADD phred score.

gnomadAFR: Numeric. Allele frequency in the African/African-American population in gnomAD.

gnomadAMR: Numeric. Allele frequency in the Latino/Admixed American population in gnomAD.

gnomadASJ: Numeric. Allele frequency in the Ashkenazi Jewish population in gnomAD.

gnomadEAS: Numeric. Allele frequency in the East Asian population in gnomAD.

gnomadFIN: Numeric. Allele frequency in the Finnish population in gnomAD.

gnomadNFE: Numeric. Allele frequency in the Non-Finnish European population in gnomAD.

gnomadNFEEST: Numeric. Allele frequency in the Estonian population in gnomAD.

gnomadNFENWE: Numeric. Allele frequency in the Northwest European population in gnomAD.

gnomadNFESEU: Numeric. Allele frequency in the Southern European population in gnomAD.

gnomadNFEONF: Numeric. Allele frequency in the Other Non-Finnish European population in gnomAD.

gnomadOTH: Numeric. Allele frequency in other populations in gnomAD.

References

https://gnomad.broadinstitute.org/

Examples

if (FALSE) { # \dontrun{
result <- variantInfo(variant_id = "rs2494663")
} # }